requirements.txt
python-telegram-bot==13.0
crochet==1.12.0
Twisted==20.3.0
Scrapy==2.4.0
